backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.0
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
bcrypt==4.0.1
bitarray==2.6.2
bitmath==1.3.3.1
blurhash==1.1.4
boto3==1.26.50
botocore==1.29.50
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cytoolz==0.12.1
dnspython==2.2.1
email-validator==1.3.0
eth-abi==4.2.1
eth-account==0.9.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.1
Flask>=2.3.0,<3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended>=4.5.0,<4.6.0
flask-openapi3==2.3.2
Flask-SQLAlchemy==3.0.2
fpdf2==2.6.1
frozenlist==1.3.3
greenlet==2.0.1
gunicorn==20.1.0
hexbytes==0.3.0
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.1
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
lru-dict==1.1.8
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
numpy==1.24.2
packaging==22.0
parsimonious==0.9.0
Pillow==9.4.0
protobuf==4.24.3
psycopg2-binary==2.9.5
py-automapper==1.2.3
pycryptodome==3.16.0
pydantic>=1.10.2,<2.0.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-http-client==3.3.7
qrcode==7.4.2
random-password-generator==2.2.0
requests==2.28.1
rlp==3.0.0
s3transfer==0.6.0
sendgrid==6.9.7
six==1.16.0
SQLAlchemy==1.4.45
starkbank-ecdsa==2.2.0
toolz==0.12.0
typing_extensions>=4.12.2
urllib3==1.26.13
varint==1.0.2
web3==6.10.0
websockets==11.0.3
Werkzeug>=2.3.0,<3.0.0
yarl==1.8.2
zipp==3.11.0